[refactor] image 유스케이스 패키지 분리
