+++ mobile/app/src/main/java/com/lacuc/pets/ui/manage/group/gallery/save/SaveImageViewModel.kt
@@ -4,7 +4,7 @@ import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
 import com.lacuc.pets.data.Result
-import com.lacuc.pets.domain.group.UploadImageUseCase
+import com.lacuc.pets.domain.image.UploadImageUseCase
 import com.lacuc.pets.util.SingleLiveEvent
 import com.lacuc.pets.util.safeValue
 import kotlinx.coroutines.launch
